More attempts at creating a post
I have fixed a couple of errors, but I am now stuck and giving up for
the night. When creating a post, I don't know how to retrieve the
UserProfile of the currently logged in user. Looking at some other code
I figured out how to get the username, but not the actual UserProfile
object associated with that username. That is not the problem though. To
see the current problem, go to /Post/Create and an error will pop up
about an invalid column.