Trying to fix things, no permissions
I apparently do not have permissions to delete a key in the database.

Cannot drop the index 'dbo.Posts.IX_user_UserId', because it does not
exist or you do not have permission.
This error shows up when I try to update the database. It has been
happening all day and so there has been no progress made on creating
posts.