Post work, Likes, Home page, still need new table
Worked a lot on the home page, posting, and liking functionality. The
current bug is that the database does not store a list of anything very
well so there is no tracking of who has liked a post. Therefore, the
counter will always be updated even if the same person keeps clicking
like. I will create a new table in the database that contains all the
appropriate pairs of foreign keys of liked posts and liking users to
solve this.