Finished post and follow functionality
Note that the lists of the user's followed posts and user's liked posts
still need to show up on the home page. I have some code commented out
in the Home/Index.cshtml page, but it does not correctly check who the
current user is. Once I can check who the current user is, the lists
should work just fine.